version to test
git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/orocos-components@6408 ce417995-dfc9-0310-95a0-acdaff106893

+++ hardware/laserscanner/LaserScanner.hpp
@@ -56,7 +56,7 @@ namespace OCL {
     protected:
         RTT::Property<int> _port;
         RTT::Property<int> _range_mode;
-        RTT::Property<int> _res_mode;
+        RTT::Property<double> _res_mode;
         RTT::Property<std::string> _unit_mode;
 
         /// Dataport which contains the measurements
